Moved add condition UI to expression add form
