+++ src/Plugin/RulesExpression/RulesAction.php
@@ -25,7 +25,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *
  * @RulesExpression(
  *   id = "rules_action",
- *   label = @Translation("An executable action."),
+ *   label = @Translation("Action"),
  *   form_class = "\Drupal\rules\Form\Expression\ActionForm"
  * )
  */
@@ -140,8 +140,11 @@ class RulesAction extends ExpressionBase implements ContainerFactoryPluginInterf
    * {@inheritdoc}
    */
   public function getLabel() {
-    $definition = $this->actionManager->getDefinition($this->configuration['action_id']);
-    return $this->t('Action: @label', ['@label' => $definition['label']]);
+    if (!empty($this->configuration['action_id'])) {
+      $definition = $this->actionManager->getDefinition($this->configuration['action_id']);
+      return $this->t('Action: @label', ['@label' => $definition['label']]);
+    }
+    return parent::getLabel();
   }
 
   /**
